Fix compilation for systems without math.h header
cbortojson and cborpretty depends on decode_half function from
compilersupport_p.h that uses math.h functions. As cbortojson and
cborpretty are usually important only in development phase, move math
dependency to another header so it is now possible to compile tinycbor
in systems that doesn't provide math.h header
